Rename Application project and update namespaces
Rename the Application project/folder to TournamentPlatformSystemWebApi.Application and update related identifiers across the solution. Updated namespaces for DTOs and interfaces, adjusted using directives in controllers, Swagger examples, tests and repository comments, and updated ProjectReference entries in API/Infrastructure csproj files and the solution file. Also modified the Dockerfile COPY path to match the new project name.
