+++ server/TournamentPlatformSystemWebApi.API/Swagger/Examples/LoginRequestExample.cs
@@ -1,5 +1,5 @@
 using Swashbuckle.AspNetCore.Filters;
-using TournamentPlatformSystem.Application.DTOs.Auth;
+using TournamentPlatformSystemWebApi.Application.DTOs.Auth;
 
 namespace TournamentPlatformSystemWebApi.API.Swagger.Examples
 {
