+++ server/TournamentPlatformSystemWebApi.API/TournamentPlatformSystemWebApi.API.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../TournamentPlatformSystem.Application/TournamentPlatformSystemWebApi.Application.csproj" />
+    <ProjectReference Include="../TournamentPlatformSystemWebApi.Application/TournamentPlatformSystemWebApi.Application.csproj" />
     <ProjectReference Include="../TournamentPlatformSystemWebApi.Core/TournamentPlatformSystemWebApi.Core.csproj" />
     <ProjectReference Include="../TournamentPlatformSystemWebApi.Infrastructure/TournamentPlatformSystemWebApi.Infrastructure.csproj" />
     <ProjectReference Include="../TournamentPlatformSystemWebApi.Common/TournamentPlatformSystemWebApi.Common.csproj" />
