+++ server/TournamentPlatformSystemWebApi.Application/DTOs/Auth/UserDto.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace TournamentPlatformSystem.Application.DTOs.Auth
+namespace TournamentPlatformSystemWebApi.Application.DTOs.Auth
 {
     public class UserDto
     {
