+++ server/TournamentPlatformSystemWebApi.Application/DTOs/MatchDto.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace TournamentPlatformSystem.Application.DTOs
+namespace TournamentPlatformSystemWebApi.Application.DTOs
 {
     public class MatchDto
     {
