+++ server/TournamentPlatformSystemWebApi.Application/DTOs/TournamentCreateDto.cs
@@ -1,11 +1,11 @@
 using System;
 
-namespace TournamentPlatformSystem.Application.DTOs
+namespace TournamentPlatformSystemWebApi.Application.DTOs
 {
-    public class TournamentDto
+    public class TournamentCreateDto
     {
-        public Guid Id { get; set; }
         public string? Title { get; set; }
+
         public string? Description { get; set; }
         public string? Conditions { get; set; }
         public DateTime StartDate { get; set; }
@@ -13,7 +13,5 @@ namespace TournamentPlatformSystem.Application.DTOs
         public DateTime RegistrationCloseDate { get; set; }
         public Guid? Sport { get; set; }
         public int MaxParticipants { get; set; }
-        public string? Status { get; set; }
-        public Guid OrganizerId { get; set; }
     }
 }
