+++ server/TournamentPlatformSystemWebApi.Application/Interfaces/IDbStateChecher.cs
