+++ server/TournamentPlatformSystemWebApi.Application/TournamentPlatformSystemWebApi.Application.csproj
